Backend/requirements.txt
anyio==3.7.0
bson==0.5.10
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
dotenv-python==0.0.1
exceptiongroup==1.1.1
fastapi==0.98.0
h11==0.14.0
idna==3.4
motor==3.2.0
pydantic==1.10.9
pymongo==4.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.3
uvicorn==0.22.0